# backend/docs/swagger.yaml
definitions:
  entity.Event:
    properties:
      description:
        type: string
      end_time:
        type: string
      event_name:
        type: string
      id:
        type: integer
      organizer_id:
        type: integer
      start_time:
        type: string
      virtual_room_id:
        type: integer
    type: object
...
  equipment_dto.CreateEquipment:
    properties:
      name:
        type: string
      user_login:
        type: string
    type: object
  event_dto.CreateEvent:
    properties:
      description:
        type: string
      end_date:
        type: string
      organizer_login:
        type: string
      start_date:
        type: string
      title:
        type: string
      virtual_room_name:
        type: string
    type: object
  event_participiant_dto.BecomeParticipantDto:
    properties:
      event:
        $ref: '#/definitions/entity.Event'
      user_id:
        type: integer
    type: object
  event_participiant_dto.LeaveEventDto:
    properties:
      event:
        $ref: '#/definitions/entity.Event'
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /equipments/{name}:
    get:
      description: Fetches the user who is associated with the specified equipment
        name
      parameters:
      - description: Equipment name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get user by equipment name
      tags:
      - equipment
  /equipments/create:
    post:
      consumes:
      - application/json
      description: Creates a new equipment entry with the provided details
      parameters:
      - description: Equipment details
        in: body
        name: equipment
        required: true
        schema:
          $ref: '#/definitions/equipment_dto.CreateEquipment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new equipment item
      tags:
      - equipment
  /eventParticipants:
    delete:
      consumes:
      - application/json
      description: Allows a user to leave a specific event they are participating
        in
      parameters:
      - description: Event leave details
        in: body
        name: leaveEvent
        required: true
        schema:
          $ref: '#/definitions/event_participiant_dto.LeaveEventDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User leaves an event
      tags:
      - event_participants
    post:
      consumes:
      - application/json
      description: Allows a user to join a specific event by becoming its participant
      parameters:
      - description: Event participant details
        in: body
        name: participant
        required: true
        schema:
          $ref: '#/definitions/event_participiant_dto.BecomeParticipantDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User becomes a participant in an event
      tags:
      - event_participants
  /eventParticipants/getAllEventsByParticipant:
    get:
      description: Fetches all events for a user identified by their participant ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all events by participant ID
      tags:
      - event_participants
  /eventParticipants/getAllParticipantsByEvent:
    post:
      consumes:
      - application/json
      description: Fetches a list of users participating in a specified event
      parameters:
      - description: Event details
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/entity.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all participants of an event
      tags:
      - event_participants
  /events/{name}:
    delete:
      consumes:
      - application/json
      description: Deletes an event by its name
      parameters:
      - description: Event name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete an event
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Fetches event details by its name
      parameters:
      - description: Event name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get an event by name
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Updates details of an existing event
      parameters:
      - description: Event name
        in: path
        name: name
        required: true
        type: string
      - description: Event update details
        in: body
        name: event
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update an event
      tags:
      - events
  /events/all:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all events
      tags:
      - events
  /events/create:
    post:
      consumes:
      - application/json
      description: Creates a new event with the provided details
      parameters:
      - description: Event details
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/event_dto.CreateEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create an event
      tags:
      - events
...
swagger: "2.0"